+++ Requirements.txt
@@ -1,11 +1,7 @@
-
-from langchain_community.document_loaders import PyPDFLoader
-from langchain_text_splitters import RecursiveCharacterTextSplitter
-from sentence_transformers import SentenceTransformer
-import chromadb
-import glob
-import os
-from flask import request, jsonify, Blueprint
-import os
-from langchain_mistralai.chat_models import ChatMistralAI
-import chromadb
+langchain_community
+langchain_text_splitters
+sentence_transformers
+chromadb
+glob2
+flask
+langchain_mistralai
